Moved token verify wrapper function under 'auth' subdirectory
