+++ fmu-to-resqml/auth/tokens.py
@@ -2,11 +2,7 @@
     All functions managing tokens
 """
 
-import jwt
-import json
 import requests
-
-from time import time
 from os import environ
 
 from flask import Request
